Version 0.13.4
Add new Enchantment system (and all data, except red)
Add poison effect system

Bug fix : Red Weapon type 1 was lvl 100
Bug fix : The writing of quality label was Orange on Costume Page
Bug fix : Possible crash

+++ GFBuilder/src/fr/vlik/grandfantasia/stats/Effect.java
@@ -124,7 +124,7 @@ public class Effect implements Calculable {
 		this.value = (int) (this.value * coef);
 	}
 	
-	public void addEnchantValue(int value) {
+	public void addEnchantValue(double value) {
 		this.value = this.value + value;
 	}
